+++ HomeReadingLibraryWeb/ClientApp/src/app/entities/missing-checkin-report-item.ts
@@ -1,0 +1,12 @@
+export interface MissingCheckinReportItem {
+  studentFirstName: string;
+  studentLastName: string;
+  studentBarCode: string;
+  teacherName: string;
+  grade: string;
+  bookTitle: string;
+  bookCopyBarCode: string;
+  readingLevel: string;
+  boxNumber: string;
+  checkedOutDate: string;
+}